Rearranged EyePose a bit due to some editor issues
