+++ Assets/Game/Users/Nap/Scripts/EyePose.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: d9ebb6e71c6d46d58556359d37006fa1
+timeCreated: 1737772035
