+++ Assets/Game/Users/Nap/Scripts/EyePoseRegistry.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: c5aeccf8e5784e23b97b5873e67fba8b
+timeCreated: 1737772285
