--- Assets/Game/Users/Nap/Scripts/EyePoses.cs.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 002b50b80ef146cf896582e8ef28d6c5
-timeCreated: 1737740296
